requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.2
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
neo4j>=5.14.1
azure-storage-blob>=12.19.0
spacy>=3.7.2
pyyaml>=6.0.1
ruamel.yaml>=0.18.5
transformers>=4.35.2
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
aiohttp>=3.9.1
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1 